Deprecated bp_core_add_nav_item() and bp_core_add_subnav_item(). Replaced with bp_core_new_nav_item() and bp_core_new_subnav_item() that support ordering of menu items, more easily understandable parameters and associated parameter array support.
Added new groups API that works very similarly to the new widgets API. More documentation on using this coming shortly.

Reformatted most of the core group code so it is easier to follow and more readable.

Completely re-wrote the group creation process to allow the insertion of custom creation steps at any point in the process. The new groups API will make it very easy to make use of this new functionality.

Updated all components to remove use of deprecated functions.